+++ src/face-detection/demo/index.js
@@ -1,0 +1,19 @@
+Page({
+  onFaceStatusChange(data, context) {
+    console.log('data', data);
+    return new Promise((resolve, reject) => {
+      context.doLeftFaceCheck().then(() => {
+        context.doRightFaceCheck().then(() => {
+          resolve();
+        }).catch(() => {
+          reject();
+        });
+      }).catch(() => {
+        reject();
+      });
+    });
+  },
+  onFail(error) {
+    console.log('error', error);
+  },
+});