+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+# 0.3.2
+
+`2018-08-07`
+
+- **Feature**
+  - 新增`popup`弹出菜单组件
+  - `face-detection`组件新增活体检测功能
+
+# 0.3.1
+
+`2018-07-27`
+
+- **Feature**
+  - `face-detection`组件新增`appName`和`serviceName`字段
+
+## 0.3.0
+
+`2018-07-26`
+
+- **Feature**
+  - 新增`face-detection`组件
+  - 新增`footer`组件
+  - `page-result`组件增加slot，方便开发者个性化定制区域内容
+
+- **Enhancement**
+  - 优化`calendar`组件在初次渲染时的闪烁问题
+  - 优化`swipe-action`右侧按钮宽度自适应文本内容
+
+
 ## 0.2.0
 
 `2018-07-11`